Series resistors on MOSI and SCK (v03)
Limits current into OLED I/O lines. This in conjunction with some code
changes will help to protect the display from overvoltage.

This required removing pull-down resistor to RESET and pull-up to SS.
BUT! You can still program the display, using an ICSP, with the series
resistors on MOSI and SCK.